restored remove function moved get_env_var to utils
