+++ Dockerfile
@@ -14,6 +14,7 @@ ENV ENABLE_SLASH_COMMANDS=False
 ENV VC_TIMOUT_DEFAULT=True
 ENV MAX_HISTORY_LENGTH=10
 ENV MAX_TRACKNAME_HISTORY_LENGTH=15
+ENV HEROKU=0
 
 RUN pip --no-cache-dir install -r requirements.txt \
     && apt-get update \
