+++ config/setup.py
@@ -1,6 +1,20 @@
 "This file is here to automatically install the selected DB package"
+import os
 from setuptools import setup
-from config import DATABASE
+
+# database url in SQL Alchemy-supported format, must be async-compatible
+# CHANGE ONLY IF YOU KNOW WHAT YOU'RE DOING
+DATABASE = "sqlite+aiosqlite:///settings.db"
+if os.getenv("HEROKU") and os.getenv("HEROKU") != "0":
+    # example with Heroku Postgres
+    DATABASE = os.getenv(
+        "DATABASE_URL",  # environment variable with the DB url
+        "postgres",  # default url (as env vars are not available at build time)
+    ).replace(
+        "postgres", "postgresql+asyncpg", 1  # make url supported by SQL Alchemy
+    )
+    # another example with MySQL
+    # DATABASE = os.getenv("DATABASE_URL", "mysql").replace("mysql", "mysql+aiomysql", 1)
 
 
 setup(
